+++ examples/ethercat/ecat_load_firmware.py
@@ -3,13 +3,13 @@ import argparse
 from ingenialink.ethercat.network import EthercatNetwork
 
 
-def ecat_load_fw(args):
+def ecat_load_fw(args: argparse.Namespace) -> None:
     net = EthercatNetwork(args.interface)
     boot_in_app = args.firmware_path.endswith(".sfu")
     net.load_firmware(args.firmware_path, boot_in_app, slave_id=args.slave_id)
 
 
-def setup_command():
+def setup_command() -> argparse.Namespace:
     parser = argparse.ArgumentParser(description="EtherCAT connection example script.")
     interface_help = """Network adapter interface name. To find it: \n
     - On Windows, \\Device\\NPF_{id}. To get the id, run the command: wmic nic get name, guid \n